moved to recursive command loading
