+++ src/commands/rips/ripcount.js
@@ -1,6 +1,6 @@
 const { SlashCommandBuilder } = require('discord.js');
 
-const rips = require('../modules/rips');
+const rips = require('../../modules/rips');
 
 module.exports = {
   data: new SlashCommandBuilder().setName('rips').setDescription('Replies with the amount of saved rips.'),
